shuffle the functions of the math folder
